+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - `FlyoutListTile` can be used outside of a flyout ([#650](https://github.com/bdlukaa/fluent_ui/issues/650))
 - Add uk localization ([#647](https://github.com/bdlukaa/fluent_ui/pull/647))
+- Add `Flyout.navigatorKey` ([#538](https://github.com/bdlukaa/fluent_ui/issues/538))
 
 ## 4.1.2